server/requirements.txt
fastapi
uvicorn
python-socketio
openai==0.28